before test & checks
