+++ wordpress/kredeum-nfts-swarm.php_handelbars
@@ -2,7 +2,7 @@
 /**
  *
  * Plugin Name: Kredeum NFTs
- * Description: Plugin to archive your medias to IPFS and mint them as NFTs
+ * Description: Plugin to archive your medias to SWARM and mint them as NFTs
  * Plugin URI: https://docs.kredeum.tech
  * Version: {{version.latest}}
  * Author: Kredeum <alain@kredeum.com>
@@ -18,6 +18,7 @@
 namespace KredeumNFTs;
 
 define( 'KREDEUM_NFTS_VERSION', '{{version.latest}}' );
+define( 'STORAGE', '{{storage}}' );
 
 defined( 'ABSPATH' ) || die( esc_html( __( 'Not allowed', 'kredeum-nfts' ) ) );
 define( 'SWARMPRESS_NFTS_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );
@@ -25,19 +26,23 @@ define( 'SWARMPRESS_NFTS_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );
 if ( is_admin() ) {
 	define( 'AUTO_ARCHIVE', get_option( 'AUTO_ARCHIVE', array( '' ) )[0] );
 
+	//require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/{{storage}}/index.php';
+
 	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/ajax/ajax.php';
+	
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/storage/query.php';
 
 	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/swarm-bee.php';
 	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/query.php';
 
-	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/media/post.php';
-	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/media/upload.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/media/post.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/media/upload.php';
 
-	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/media-list/actions.php';
-	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/media-list/column.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/media-list/actions.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/media-list/column.php';
 
-	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/settings/class-settings.php';
-	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/settings/fields.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/settings/class-settings.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/settings/fields.php';
 
 	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'vendor/tcdent/php-restclient/restclient.php';
 
@@ -56,15 +61,18 @@ if ( is_admin() ) {
 	);
 }
 
-define( 'SWARM_SERVER', 'https://api.gateway.ethswarm.org' );
-define( 'SWARM_GATEWAY', 'https://api.gateway.ethswarm.org/bzz/' );
+define( 'SWARM_SERVER', '{{swarmServer}}' );
+define( 'STORAGE_GATEWAY', '{{swarmGateway}}' );
 define( 'SWARM_NODE_URL', get_option( 'SWARM_NODE_URL', '' ) ? get_option( 'SWARM_NODE_URL', '' ) : SWARM_SERVER );
 define( 'SWARM_BATCH_ID', get_option( 'SWARM_BATCH_ID', '' ) ? get_option( 'SWARM_BATCH_ID', '' ) : '0000000000000000000000000000000000000000000000000000000000000000' );
 
-require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/swarm/cid.php';
-require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/swarm/link.php';
+require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/storage/cid.php';
+require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/storage/link.php';
 
 require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/shortcode/shortcode.php';
+
+require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/helpers/storage.php';
+
 
 add_action(
 	'plugins_loaded',
