+++ wordpress/kredeum-nfts.php_handelbars
@@ -2,7 +2,7 @@
 /**
  *
  * Plugin Name: Kredeum NFTs
- * Description: Plugin to archive your medias to IPFS and mint them as NFTs
+ * Description: Plugin to archive your medias to SWARM and mint them as NFTs
  * Plugin URI: https://docs.kredeum.tech
  * Version: {{version.latest}}
  * Author: Kredeum <alain@kredeum.com>
@@ -21,69 +21,58 @@ define( 'KREDEUM_NFTS_VERSION', '{{version.latest}}' );
 define( 'STORAGE', '{{storage}}' );
 
 defined( 'ABSPATH' ) || die( esc_html( __( 'Not allowed', 'kredeum-nfts' ) ) );
-define( 'KREDEUM_NFTS_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );
+define( 'SWARMPRESS_NFTS_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );
 
 if ( is_admin() ) {
-	define( 'IPFS_CID_VERSION', '1' );
 	define( 'AUTO_ARCHIVE', get_option( 'AUTO_ARCHIVE', array( '' ) )[0] );
-	define( 'KREDEUM_BETA', get_option( 'KREDEUM_BETA', array( '' ) )[0] );
-	define( 'IPFS_API', get_option( 'IPFS_API', array( '' ) )[0] );
-	define( 'IPFS_CLUSTER_API', get_option( 'IPFS_CLUSTER_API', array( '' ) )[0] );
-	define( 'IPFS_PINNING_API', get_option( 'IPFS_PINNING_API', array( '' ) )[0] );
 
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/nfts/index.php';
+	//require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/{{storage}}/index.php';
 
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/ajax/ajax.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/ajax/ajax.php';
+	
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/storage/query.php';
 
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/storage/query.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/swarm-bee.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/swarm/query.php';
 
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/ipfs/import.php';
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/ipfs/multipart.php';
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/ipfs/nft-storage.php';
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/ipfs/query.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/media/post.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/media/upload.php';
 
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/media/post.php';
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/media/upload.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/media-list/actions.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/media-list/column.php';
 
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/media-list/actions.php';
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/media-list/column.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/settings/class-settings.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'admin/settings/fields.php';
 
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/settings/class-settings.php';
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'admin/settings/fields.php';
-
-	require_once KREDEUM_NFTS_PLUGIN_PATH . 'vendor/tcdent/php-restclient/restclient.php';
+	require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'vendor/tcdent/php-restclient/restclient.php';
 
 	$api = new \RestClient( array( 'base_url' => '' ) );
 
 	add_action(
 		'admin_enqueue_scripts',
 		function ( $hook ) {
-			if ( 'nfts_page_ipfs_settings' === $hook || 'toplevel_page_nfts' === $hook || 'upload.php' === $hook ) {
+			if ( 'toplevel_page_swarm_settings' === $hook || 'upload.php' === $hook ) {
 				wp_enqueue_script( 'kredeum_nfts', plugin_dir_url( __FILE__ ) . 'lib/js/kredeum-nfts.js', array(), KREDEUM_NFTS_VERSION, true );
 				wp_register_style( 'kredeum_nfts_css', plugin_dir_url( __FILE__ ) . 'lib/js/kredeum-nfts.css', KREDEUM_NFTS_VERSION, true );
 				wp_enqueue_style( 'kredeum_nfts_css' );
-			}
-			if ( 'toplevel_page_nfts' === $hook ) {
-				wp_register_style( 'kredeum_nfts_front_css', plugin_dir_url( __FILE__ ) . 'lib/css/front.css', KREDEUM_NFTS_VERSION, true );
-				wp_enqueue_style( 'kredeum_nfts_front_css' );
 			}
 		},
 		110
 	);
 }
 
-define( 'IPFS_GATEWAY', 'https://ipfs.io/ipfs/' );
-define( 'NFT_STORAGE_KEY', get_option( 'NFT_STORAGE_KEY', '' ) ? get_option( 'NFT_STORAGE_KEY', '' ) : 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJkaWQ6ZXRocjoweEVkYzMzMjZEODc3OGU5QWVhMGZGMzc0MmEyRERGRDREM0E0NTYzNTciLCJpc3MiOiJuZnQtc3RvcmFnZSIsImlhdCI6MTYyMTAwMjI3ODIzNywibmFtZSI6Im9wZW4tbmZ0cyJ9.EJeI37e2D9k09coXLIz8IgqFi85gHlLteE-ETMcRt8I' );
+define( 'SWARM_SERVER', '{{swarmServer}}' );
+define( 'STORAGE_GATEWAY', '{{swarmGateway}}' );
+define( 'SWARM_NODE_URL', get_option( 'SWARM_NODE_URL', '' ) ? get_option( 'SWARM_NODE_URL', '' ) : SWARM_SERVER );
+define( 'SWARM_BATCH_ID', get_option( 'SWARM_BATCH_ID', '' ) ? get_option( 'SWARM_BATCH_ID', '' ) : '0000000000000000000000000000000000000000000000000000000000000000' );
 
-require_once KREDEUM_NFTS_PLUGIN_PATH . 'common/ipfs/cid.php';
-require_once KREDEUM_NFTS_PLUGIN_PATH . 'common/ipfs/link.php';
-require_once KREDEUM_NFTS_PLUGIN_PATH . 'common/ipfs/links.php';
+require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/storage/cid.php';
+require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/storage/link.php';
 
-require_once KREDEUM_NFTS_PLUGIN_PATH . 'common/helpers/storage.php';
+require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/shortcode/shortcode.php';
 
-require_once KREDEUM_NFTS_PLUGIN_PATH . 'common/shortcode/shortcode.php';
+require_once SWARMPRESS_NFTS_PLUGIN_PATH . 'common/helpers/storage.php';
 
-// require_once KREDEUM_NFTS_PLUGIN_PATH . 'public/post/bottom.php'; .
 
 add_action(
 	'plugins_loaded',
