+++ wordpress/plugins/kredeum-nfts/admin/ipfs/import.php
@@ -5,7 +5,7 @@
  * @package kredeum/nfts
  */
 
-namespace KredeumNFTs\Ipfs;
+namespace KredeumNFTs\Storage\Ipfs;
 
 /**
  * IPFS import url
