+++ wordpress/plugins/kredeum-nfts/admin/ipfs/index.php
