+++ wordpress/plugins/kredeum-nfts/admin/ipfs/multipart.php
@@ -5,7 +5,7 @@
  * @package kredeum/nfts
  */
 
-namespace KredeumNFTs\Ipfs;
+namespace KredeumNFTs\Storage\Ipfs;
 
 /**
  * Multipart
