+++ wordpress/plugins/kredeum-nfts/admin/ipfs/query.php
@@ -7,23 +7,6 @@
 
 namespace KredeumNFTs\Storage\Ipfs;
 
-use function \KredeumNFTs\Storage\getStorageRef;
-
-/**
- * IPFS upsert file
- *
- * @param string $post_id Id post.
- *
- * @return string CID hash
- */
-function upsert( $post_id ) {
-	$file = get_attached_file_meta( $post_id );
-	if ( '' === $file->cid ) {
-		$cid = insert( $post_id, $file->filename );
-	}
-	return $cid;
-}
-
 /**
  * IPFS insert file
  *
@@ -35,6 +18,5 @@ function insert( $post_id ) {
 	if ( defined( 'NFT_STORAGE_KEY' ) ) {
 		$cid = nft_storage_add_and_pin( $post_id );
 	}
-	update_post_meta( $post_id, getStorageRef(), $cid );
 	return $cid;
 }
