+++ wordpress/plugins/kredeum-nfts/admin/media/upload.php
@@ -5,7 +5,7 @@
  * @package kredeum/nfts
  */
 
-namespace KredeumNFTs\Ipfs;
+namespace KredeumNFTs\Storage;
 
 /**
  * IPFS new action
