--- wordpress/plugins/kredeum-nfts/common/ipfs/link.php
@@ -1,35 +1,0 @@
-<?php
-/**
- * Public IPFS LINK function
- *
- * @package kredeum/nfts
- */
-
-namespace KredeumNFTs\Ipfs;
-
-/**
- * Return IPFS url
- *
- * @param string $cid file CID.
- * @return string cid with path
- */
-function url( $cid ) {
-	return $cid ? IPFS_GATEWAY . $cid : '';
-}
-
-/**
- * Return IPFS link
- *
- * @param string $cid : file CID.
- * @param string $text : text for the link.
- * @return string html link to CID with text
- */
-function link( $cid, $text = '' ) {
-	if ( ! $text ) {
-		$text = $cid;
-	}
-	$url = esc_url( url( $cid ) );
-	$txt = esc_html( $text );
-
-	return $cid ? "<a href='$url'>$txt</a>" : '';
-}
