+++ wordpress/plugins/kredeum-nfts/common/ipfs/links.php
@@ -5,7 +5,7 @@
  * @package kredeum/nfts
  */
 
-namespace KredeumNFTs\Ipfs;
+namespace KredeumNFTs\Storage\Ipfs;
 
 /**
  * List IPFS links
